requirements.txt
flask==2.3.3
tensorflow==2.21.0
Pillow==10.0.0
numpy>=1.26.0
werkzeug==2.3.7
gunicorn==21.2.0
